Extract useExplorerAddressLink.ts
